backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.9
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
identify==2.6.9
idna==3.10
mypy==1.15.0
mypy-extensions==1.0.0
nodeenv==1.9.1
platformdirs==4.3.7
pre_commit==4.2.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
types-requests==2.32.0.20250328
typing_extensions==4.13.1
urllib3==2.3.0
virtualenv==20.30.0
